requirements.txt
PySide6
mss
sounddevice
soundfile
numpy
python-slugify
pyyaml
